Gigacommit with refactoring
WIP

Refactor of logic

Refractoring of mytoninstaller.py

Fixed inject_globals

Removed print

Fixed typos

WIP

WIP

WIP

WIP

WIP2

WIP

RC

Fix

Fixed scripts

Crc16 -> fastcrc

tonhttpapi installer

Fixed merge

WIP

Split on three modules, added mypylib and mypyconsole submodules

Added imports in mytoncore

Updated mtc-jsonrpc installer

update mtc-jsonrpc installer

WIP

Try fix merge conflicts

Rebase. New commits apply

1.  a61ab21 - add ubuntu 22.04 to the list of tested OS
2.  a41ad04 - Update README.Ru.md
3.  b12f234 - Add Traditional Chinese Support and Fix JSON Format Errors
4.  d152a70 - Update README with tested operating systems and installation scripts
5.  329c6bd - Refactor wallet import documentation
6.  5d6afcd - feat: Add import wallet documentation
7.  d00a94a - Refactor: Rename FAQ files to markdown format
8.  e417285 - Add import-wallets.md file for importing wallets with private keys or mnemonic phrases.
9.  17b45b1 - Refactor directory usage and add error handling
10. 2458f8b - Add FAQ.md with instructions on how to install, remove and configure MyTonCtrl
11. 3748e9c - Refactor MyTonCtrl directory usage
12. dd5825d - Add instructions for using mytonctrl on Ubuntu
13. 559480d - Refactor installation instructions and add operability test
14. 4e41c94 - Refactor installation instructions for mytonctrl on Ubuntu
15. a9df737 - Refactor pool activation and deposit instructions
16. 4fb4a0f - Add Chinese translation for Nominator Pool documentation.
17. 47ad5cf - feat: Add instructions for running a validator node in nominator pool mode
18. c36fa7f - Refactor documentation for running a validator in nominator pool
19. b9c068f - Refactor Nominator Pool documentation for clarity
20. 95f6823 - Refactor README.md to include more detailed information
21. e52aa1c - Merge branch 'master' into update-documentation
22. f03095f - Add badges and fix formatting in README
22. cc711df - Refactor documentation structure and add MyTonCtrl installer mode
23. 4294fa8 - Refactor commit message:
24. 9d25122 - Update FAQ.md with improved formatting and content
25. 6f759c1 - Update Chinese translation of nominator-pool.md
26. 8c17c60 - Add stake param setting for nominator pool docs
27. 776cde0 - Fix tabs
28. 456b2dc - Merge remote-tracking branch 'sonofmom/feature/update_docs' into update-documentation
29. 9a993a7 - Add null stake option to usePool command
30. 2556501 - upgrade with ninja
31. ec3a8b9 - persist global.config.json before the upgrade
32. d51d20e - make persistence of global.config.json rights friendly
33. b3bc22e - clean up build directory (without removing it) prior to upgrade
34. 953e991 - avoid usage of /tmp dir on clean up
35. 8e14dec - Merge pull request #127 from Gusarich/master
36. e5978dd - Merge pull request #128 from awesome-doge/add-traditional-chinese-support
37. 4e97349 - Merge pull request #129 from awesome-doge/update-documentation
38. c1eee8d - Merge pull request #131 from neodiX42/upgrade-with-ninja
39.

Applying commit efbf35c "update mypylib"

Replay commit ef7d665 "bugfix"

Replay commit 8733452 "upgrade bugfix"

Replay commit 2a359a8 "update mypylib"

Replay commits:
1. f1f5247 - add GetDomainFromAuction function
2. ac0d138 - Merge pull request #135 from ton-blockchain/master
3. 17fd02c - Merge pull request #136 from ton-blockchain/dev

Replay commit dc56e83 "get_service_uptime bugfix on systems without systemd"

Replay commits:
1. c3f3119 - mytoninstaller.py bugfix
2. 2c68e2f - bugfix if the files are in the wrong place

Replay commit c789ebc "bugfix and moving to PEP8 stile"

Replay commit 857b5a5 "bugfix"

Replay commit 966f1a7 "bugfix" and update install.sh

Refactor installation instructions for mytonctrl on Ubuntu

The commit refactors the installation instructions for mytonctrl on Ubuntu. The changes include updating the language, adding numbered steps, and clarifying commands. Additionally, the commit adds a section to check wallet balances and activate wallets. Finally, it includes information about checking logs and using help commands.

Refactor documentation for running a validator in nominator pool

The commit message summarizes the changes made to the documentation for running a validator in a nominator pool. The changes include updating instructions on hardware requirements, recommended providers, and steps for creating and activating pools. Additionally, the commit includes instructions for switching from an individual validator to a nominator pool.

Refactor Nominator Pool documentation for clarity

This commit refactors the Nominator Pool documentation to improve its readability and clarity. Changes include rewording sentences, adding hyperlinks, and providing additional explanations for certain steps. The commit also includes a section on transitioning a regular validator to nominator pool mode.

Refactor README.md to include more detailed information

This commit refactors the README.md file to include a table of contents, installation instructions, and useful links. It also adds documentation in English, Russian, and Mandarin for frequently asked questions, importing wallets, manuals for Ubuntu, and nominator pools. Additionally, it includes information on telemetry and a web admin panel.

Refactor documentation structure and add MyTonCtrl installer mode

This commit refactors the documentation structure for MyTonCtrl, categorizing technical documents by language. It also adds a new feature, MyTonCtrl installer mode, which includes a web admin panel and local copy of toncenter. The web admin panel allows users to control the node/validator through their browser after installing an additional module. A local copy of toncenter can be set up on the user's server by enabling another module in the installer.

add GetDomainFromAuction function

Fix duplicated method GetDomainFromAuction

+++ scripts/uninstall.sh
@@ -1,4 +1,10 @@
 #!/bin/bash
+full=true
+while getopts f flag; do
+	case "${flag}" in
+		f) full=false
+	esac
+done
 
 # Проверить sudo
 if [ "$(id -u)" != "0" ]; then
@@ -26,11 +32,17 @@ rm -rf /etc/systemd/system/dht-server.service
 systemctl daemon-reload
 
 # Удаление файлов
-rm -rf /usr/src/ton
+if $full; then
+	echo "removing Ton node"
+	rm -rf /usr/src/ton
+	rm -rf /usr/bin/ton
+	rm -rf /var/ton-work
+	rm -rf /var/ton-dht-server
+fi
+
 rm -rf /usr/src/mytonctrl
-rm -rf /usr/bin/ton
-rm -rf /var/ton-work
-rm -rf /var/ton-dht-server
+rm -rf /usr/src/mtc-jsonrpc
+rm -rf /usr/src/pytonv3
 rm -rf /tmp/myton*
 rm -rf /usr/local/bin/mytoninstaller/
 rm -rf /usr/local/bin/mytoncore/mytoncore.db
@@ -38,10 +50,17 @@ rm -rf /home/${user}/.local/share/mytonctrl
 rm -rf /home/${user}/.local/share/mytoncore/mytoncore.db
 
 # Удаление ссылок
-rm -rf /usr/bin/fift
-rm -rf /usr/bin/liteclient
-rm -rf /usr/bin/validator-console
+if $full; then
+	echo "removing ton node"
+	rm -rf /usr/bin/fift
+	rm -rf /usr/bin/liteclient
+	rm -rf /usr/bin/validator-console
+fi
 rm -rf /usr/bin/mytonctrl
+
+# removing pip packages
+pip3 uninstall -y myton mypylib mypyconsole
+pip3 uninstall -y ton-http-api
 
 # Конец
 echo -e "${COLOR}Uninstall Complete${ENDC}"
